db_api/requirements.txt
Flask==2.1.2
psycopg2==2.9.3
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10   